Interface migration to scripts - PipBoy, Timer, Use.
git-svn-id: https://cvet.by:8443/svn/FOnlineSource@847 c2c6a35e-ad2b-4e1f-bb64-97b989be360b

Former-commit-id: 1f1d917e09e4b6864838b500b883e7a47befbc07

+++ Source/EntityManager.h
@@ -11,7 +11,7 @@ class EntityManager
 {
 private:
     EntityMap allEntities;
-    uint      entitiesCount[ EntityType::Max ];
+    uint      entitiesCount[ (int) EntityType::Max ];
     Mutex     entitiesLocker;
 
 public:
